okcoin risk rate double

+++ xchange-okcoin/src/main/java/org/knowm/xchange/okcoin/dto/account/OkcoinFuturesFundsCross.java
@@ -9,14 +9,14 @@ public class OkcoinFuturesFundsCross {
   private final BigDecimal keepDeposits;
   private final BigDecimal profitReal;
   private final BigDecimal profitUnreal;
-  private final int riskRate;
+  private final double riskRate;
 
   public OkcoinFuturesFundsCross(
       @JsonProperty("account_rights") BigDecimal accountRights,
       @JsonProperty("keep_deposit") BigDecimal keepDeposits,
       @JsonProperty("profit_real") BigDecimal profitReal,
       @JsonProperty("profit_unreal") BigDecimal profitUnreal,
-      @JsonProperty("risk_rate") int riskRate) {
+      @JsonProperty("risk_rate") double riskRate) {
 
     this.accountRights = accountRights;
     this.keepDeposits = keepDeposits;
@@ -45,7 +45,7 @@ public class OkcoinFuturesFundsCross {
     return profitUnreal;
   }
 
-  public int getRiskRate() {
+  public double getRiskRate() {
 
     return riskRate;
   }
